+++ src/Svg.Skia/Svg/CircleDrawable.cs
@@ -1,0 +1,80 @@
+﻿#if !USE_PICTURE
+using SkiaSharp;
+using CropRect = SkiaSharp.SKImageFilter.CropRect;
+#endif
+#if USE_PICTURE
+using SKRect = Svg.Picture.Rect;
+#endif
+
+namespace Svg.Skia
+{
+    internal sealed class CircleDrawable : DrawablePath
+    {
+        private CircleDrawable()
+            : base()
+        {
+        }
+
+        public static CircleDrawable Create(SvgCircle svgCircle, SKRect skOwnerBounds, DrawableBase? parent, Attributes ignoreAttributes = Attributes.None)
+        {
+            var drawable = new CircleDrawable
+            {
+                Element = svgCircle,
+                Parent = parent,
+                IgnoreAttributes = ignoreAttributes
+            };
+
+            drawable.IsDrawable = drawable.CanDraw(svgCircle, drawable.IgnoreAttributes) && drawable.HasFeatures(svgCircle, drawable.IgnoreAttributes);
+
+            if (!drawable.IsDrawable)
+            {
+                return drawable;
+            }
+
+            drawable.Path = svgCircle.ToSKPath(svgCircle.FillRule, skOwnerBounds, drawable.Disposable);
+            if (drawable.Path is null || drawable.Path.IsEmpty)
+            {
+                drawable.IsDrawable = false;
+                return drawable;
+            }
+
+            drawable.IsAntialias = SvgExtensions.IsAntialias(svgCircle);
+
+            drawable.TransformedBounds = drawable.Path.Bounds;
+
+            drawable.Transform = SvgExtensions.ToSKMatrix(svgCircle.Transforms);
+
+            bool canDrawFill = true;
+            bool canDrawStroke = true;
+
+            if (SvgExtensions.IsValidFill(svgCircle))
+            {
+                drawable.Fill = SvgExtensions.GetFillSKPaint(svgCircle, drawable.TransformedBounds, ignoreAttributes, drawable.Disposable);
+                if (drawable.Fill is null)
+                {
+                    canDrawFill = false;
+                }
+            }
+
+            if (SvgExtensions.IsValidStroke(svgCircle, drawable.TransformedBounds))
+            {
+                drawable.Stroke = SvgExtensions.GetStrokeSKPaint(svgCircle, drawable.TransformedBounds, ignoreAttributes, drawable.Disposable);
+                if (drawable.Stroke is null)
+                {
+                    canDrawStroke = false;
+                }
+            }
+
+            if (canDrawFill && !canDrawStroke)
+            {
+                drawable.IsDrawable = false;
+                return drawable;
+            }
+
+            // TODO: Transform _skBounds using _skMatrix.
+            drawable.TransformedBounds = drawable.Transform.MapRect(drawable.TransformedBounds);
+
+            return drawable;
+        }
+    }
+}
